geonode_dominode/requirements.txt
-e git+https://github.com/GeoNode/geonode.git@3.x#egg=GeoNode

djangorestframework~=3.11
django-filter~=2.3
django-json-widget~=1.0

